requirements.txt
fastapi
uvicorn
langchain-community
langchain-openai
langgraph
llama-index
openai
matplotlib
chromadb
python-dotenv
sentence-transformers
python-socketio
langchainhub